- Created Editor Modes (MapEditor and SpriteSheetViewer) currently just placeholders. - Got switching between modes working - SpriteSheetViewer is able to show the whole spritesheet and logs what index the sprite is at in the sheet. (v1)

+++ Engine/Input/InputSystem.cpp
@@ -208,8 +208,8 @@ void InputSystem::PollEvents()
 
 		if (theEvent.type == SDL_MOUSEMOTION)
 		{
-			m_mousePosition.x = theEvent.motion.x;
-			m_mousePosition.y = theEvent.motion.y;
+			m_mousePosition.x = (float) theEvent.motion.x;
+			m_mousePosition.y = (float) theEvent.motion.y;
 		}
 	}
 }
